requirements.txt
configparser==5.0.0
neat-python-0.92
pygame==1.9.6
